chore(refactor): move EnrichedCollection schema as pivot/types.ts (refs #47)
