+++ src/cli/render-catalog.ts
@@ -1,7 +1,7 @@
 import { mkdirSync, rmSync, writeFileSync } from 'node:fs';
 import { isAbsolute, join, relative, resolve, sep } from 'node:path';
 
-import type { EnrichedCollection } from '../types';
+import type { EnrichedCollection } from '@/pivot/types';
 import { renderCollectionSchema } from '../renderers/collection-schema';
 
 export interface RenderCatalogOptions {
