+++ src/enrichment/load-enriched-collections.ts
@@ -1,4 +1,4 @@
-import type { EnrichedCollection } from '../types';
+import type { EnrichedCollection } from '@/pivot/types';
 import { loadCollectionOverwrite } from '../overwrite/overwrite-store';
 import { validateOverwriteReferences } from '../overwrite/overwrite';
 import { loadSourceCollections } from '../source/source-store';
