+++ src/pivot/types.ts
@@ -1,3 +1,4 @@
+import type { SourceCollectionProperty } from '@/source/types';
 import { z } from 'zod';
 
 /*
@@ -53,3 +54,38 @@ export function isGeometryType(value: unknown): boolean {
   }
   return (GEOMETRY_TYPES as readonly CollectionPropertyType[]).includes(value);
 }
+
+/*
+ * ============================================================================
+ * Internal Enriched Model - the pivot model
+ * ============================================================================
+ */
+
+// Canonical internal model after applying overwrites.
+
+export type EnrichedCollection = {
+  id: string;
+  namespace: string;
+  name: string;
+  title: string;
+  description: string;
+  'x-ign-theme'?: string;
+  'x-ign-selectionCriteria'?: string;
+  'x-ign-representedFeatures'?: string[];
+  required?: string[];
+  properties: EnrichedCollectionProperty[];
+};
+
+export type EnrichedCollectionProperty = SourceCollectionProperty & {
+  title?: string;
+  description?: string;
+  oneOf?: EnrichedCollectionValue[];
+};
+
+export type EnrichedCollectionValue = {
+  const: string;
+  title: string;
+  description?: string;
+  'x-ign-representedFeatures'?: string[];
+};
+
