+++ src/types.ts
@@ -1,5 +1,4 @@
 import { z } from 'zod';
-import type { SourceCollectionProperty } from './source/types';
 
 /*
  * ============================================================================
@@ -19,39 +18,6 @@ export class UnexpectedTypeError extends Error {
         this.name = 'UnexpectedTypeError';
     }
 }
-
-/*
- * ============================================================================
- * Internal Enriched Model
- * ============================================================================
- */
-
-// Canonical internal model after applying overwrites.
-export type EnrichedCollection = {
-  id: string;
-  namespace: string;
-  name: string;
-  title: string;
-  description: string;
-  'x-ign-theme'?: string;
-  'x-ign-selectionCriteria'?: string;
-  'x-ign-representedFeatures'?: string[];
-  required?: string[];
-  properties: EnrichedCollectionProperty[];
-};
-
-export type EnrichedCollectionProperty = SourceCollectionProperty & {
-  title?: string;
-  description?: string;
-  oneOf?: EnrichedCollectionValue[];
-};
-
-export type EnrichedCollectionValue = {
-  const: string;
-  title: string;
-  description?: string;
-  'x-ign-representedFeatures'?: string[];
-};
 
 /*
  * ============================================================================
