+++ test/unit/renderers/collection-schema.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import type { EnrichedCollection } from '@/types';
+import type { EnrichedCollection } from '@/pivot/types';
 import { renderCollectionSchema } from '@/renderers/collection-schema';
 
 describe('renderCollectionSchema', () => {
